move validate names function to names package
